// apps/client/src/app/model/team/webhook-setting-type.ts
export enum WebhookSettingType {
  MEMBER_JOINED = 'MEMBER_JOINED',
  MEMBER_LEFT = 'MEMBER_LEFT',

  ITEM_ADDED = 'ITEM_ADDED',
  ITEM_REMOVED = 'ITEM_REMOVED',

  LIST_ADDED = 'LIST_ADDED',
  LIST_REMOVED = 'LIST_REMOVED',

  LIST_PROGRESSION = 'LIST_PROGRESSION',
  ITEM_COMPLETION = 'ITEM_COMPLETION',
  ITEM_PROGRESSION = 'ITEM_PROGRESSION',
  FINAL_LIST_PROGRESSION = 'FINAL_LIST_PROGRESSION',
  USER_ASSIGNMENT = 'USER_ASSIGNMENT',
}